Add Settings activity, add db username option to Profile
