+++ app/build.gradle
@@ -44,4 +44,6 @@ dependencies {
 
 
 
+
+
 apply plugin: 'com.google.gms.google-services'
