--- app/src/main/java/User.java
@@ -1,19 +1,0 @@
-/**
- * Created by lordis on 1/26/2018.
- */
-
-public class User {
-
-    public String username;
-    public String email;
-
-    public User() {
-        // Default constructor required for calls to DataSnapshot.getValue(User.class)
-    }
-
-    public User(String username, String email) {
-        this.username = username;
-        this.email = email;
-    }
-
-}
