+++ app/src/main/java/org/trackhouse/trackhouse/Home.java
@@ -8,7 +8,6 @@ import android.support.design.widget.BottomNavigationView;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.MenuItem;
-import android.widget.TextView;
 
 public class Home extends AppCompatActivity {
 
@@ -42,23 +41,17 @@ public class Home extends AppCompatActivity {
                         startActivity(intent2);
                         break;
 
-                    case R.id.navigation_following:
-                        // User selects "Following" item
-                        Intent intent3 = new Intent(Home.this, Following.class);
-                        startActivity(intent3);
-                        break;
-
                     case R.id.navigation_favorites:
                         // User selects "Favorites" item
                         Intent intent4 = new Intent(Home.this, Favorites.class);
                         startActivity(intent4);
                         break;
 
-                    case R.id.navigation_profile:
-                        // User selects "Profile" item
-                        Intent intent5 = new Intent(Home.this, Profile.class);
+                    case R.id.navigation_settings:
+                        // User selects "Settings" item
+                        Intent intent5 = new Intent(Home.this, Settings.class);
                         startActivity(intent5);
-                        Log.d(TAG, "Profile navigation from Home");
+                        Log.d(TAG, "Settings navigation from Home");
                         break;
                 }
                 return false;
