+++ app/src/main/java/org/trackhouse/trackhouse/UserInformation.java
@@ -1,0 +1,17 @@
+package org.trackhouse.trackhouse;
+
+/**
+ * Created by lordis on 1/29/2018.
+ */
+
+public class UserInformation {
+
+    public String username;
+
+    public UserInformation() {
+    }
+
+    public UserInformation(String username){
+        this.username = username;
+    }
+}
