+++ app/src/main/res/layout/activity_signup.xml
@@ -37,26 +37,6 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/toolbar" />
 
-    <EditText
-        android:id="@+id/username_signup"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:layout_marginTop="48dp"
-        android:hint="@string/username_entry"
-        android:imeActionId="6"
-        android:imeActionLabel="@string/action_sign_in_short"
-        android:imeOptions="actionUnspecified"
-        android:inputType="textVisiblePassword|textNoSuggestions"
-        android:maxLines="1"
-        android:singleLine="true"
-        android:textColor="@android:color/white"
-        android:textColorHint="@color/colorAccent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/email_signup"
-        tools:layout_editor_abxoluteX="16dp" />
 
     <EditText
         android:id="@+id/password_signup"
@@ -73,7 +53,7 @@
         android:textColorHint="@color/colorAccent"
         tools:layout_editor_absoluteX="16dp"
         tools:layout_editor_absoluteY="134dp"
-        app:layout_constraintTop_toBottomOf="@id/username_signup"/>
+        app:layout_constraintTop_toBottomOf="@id/email_signup"/>
 
     <EditText
         android:id="@+id/password_reenter"
