+++ app/src/main/res/menu/navigation.xml
@@ -13,18 +13,13 @@
 
 
     <item
-        android:id="@+id/navigation_following"
-        android:icon="@drawable/ic_favorites"
-        android:title="@string/title_following" />
-
-    <item
         android:id="@+id/navigation_favorites"
         android:icon="@drawable/ic_favorite_tracks"
         android:title="@string/title_favorites" />
 
     <item
-        android:id="@+id/navigation_profile"
-        android:icon="@drawable/ic_profile"
-        android:title="@string/title_profile" />
+        android:id="@+id/navigation_settings"
+        android:icon="@drawable/ic_settings"
+        android:title="@string/title_settings" />
 
 </menu>
